move lib/dbus.py to dbus/utilities.py
Soft link back to lib/dbus.py.
Waf will install a plain file instead of a link.
No big deal, and the note should help in case of confusion.
